+++ src/main/java/id/co/mii/ta/ticketingserverside/model/dto/request/EmployeeRequest.java
@@ -19,7 +19,7 @@ public class EmployeeRequest {
    private String phoneNumber;
    private String username;
    private String password;
-   private Long roles;
+   private Integer roles;
    private Boolean isAccountLocked = false;
 
 }
