+++ src/main/java/id/co/mii/ta/ticketingserverside/repository/HistoryRepository.java
@@ -14,6 +14,6 @@ import org.springframework.stereotype.Repository;
  * @author Fathullah
  */
 @Repository
-public interface HistoryRepository extends JpaRepository<History, Long> {
+public interface HistoryRepository extends JpaRepository<History, Integer> {
     
 }
